requirements.txt
asgiref==3.7.2
Django==5.0.3
sqlparse==0.4.4
typing_extensions==4.10.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
python-dotenv==1.0.0
drf-yasg==1.21.7
validate-docbr==1.10.0
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.3.5
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13